requirements/ci.txt
-r ./base.txt
pytest-django>=3.1.2
flake8>=3.2.1
pyflakes>=1.5.0
